requirements.txt
requests
pytest
selenium
allure-pytest
